+++ src/Components/ProfilePicChange/ProfilePicChangeRemoveButton.js
@@ -1,0 +1,29 @@
+import React, { useContext } from "react";
+import { useTranslation } from "react-i18next";
+import { Button } from "../../SubComponents/Button";
+import { store } from "./ProfilePicChangeContainer";
+import { ReactComponent as XIcon } from "../../assets/x-icon.svg";
+import { REMOVE_IMAGE } from "../../utils/action-types";
+
+export const ProfilePicChangeRemoveButton = ({
+  name,
+  ...otherProps
+}) => {
+  const {
+    dispatch,
+    state: { isSubmitting }
+  } = useContext(store);
+
+  const { t } = useTranslation("userEdit");
+
+  return (
+    <Button
+      icon={<XIcon className="plc-h-6 plc-w-6" />}
+      isLoading={isSubmitting}
+      onClick={() => dispatch({ type: REMOVE_IMAGE })}
+      {...otherProps}
+    >
+      {name ?? t("labels.removeImage")}
+    </Button>
+  );
+};