+++ src/SubComponents/Modal.js
@@ -10,11 +10,11 @@ export function Modal({
   onDisplay,
   className = "",
   hideCloseButton = !window.Pelcro.paywall.displayCloseButton(),
+  hideHeaderLogo = true,
   children,
   ...props
 }) {
   const resetView = usePelcro((state) => state.resetView);
-
   useEffect(() => {
     onDisplay?.();
   }, []);
@@ -23,6 +23,8 @@ export function Modal({
     props?.onClose?.();
     resetView();
   };
+
+  const logoUrl = window.Pelcro.site.read().logo.url;
 
   return (
     <div className="pelcro-modal-overlay">
@@ -45,6 +47,15 @@ export function Modal({
                 <CloseIcon className="plc-fill-current" />
               </button>
             )}
+            <div className="plc-flex plc-items-center plc-justify-center plc-w-full">
+              <img
+                alt="business logo"
+                className={`plc-max-h-14 plc-mt-2 pelcro-modal-logo ${
+                  hideHeaderLogo ? "plc-hidden" : ""
+                }`}
+                src={logoUrl}
+              />
+            </div>
           </div>
           {children.find(({ type }) => type === ModalBody)}
           {children.find(({ type }) => type === ModalFooter)}